Refactor all errors to use errors module
