+++ src/certs/idcerttbs.rs
@@ -11,7 +11,8 @@ use x509_cert::name::Name;
 use x509_cert::serial_number::SerialNumber;
 use x509_cert::time::Validity;
 
-use crate::{Constrained, Error, IdCertToTbsCert, TbsCertToIdCert};
+use crate::errors::composite::{IdCertToTbsCert, TbsCertToIdCert};
+use crate::Constrained;
 
 use super::PublicKeyInfo;
 
@@ -57,7 +58,7 @@ pub struct IdCertTbs {
 }
 
 impl<P: Profile> TryFrom<TbsCertificateInner<P>> for IdCertTbs {
-    type Error = Error;
+    type Error = TbsCertToIdCert;
 
     fn try_from(value: TbsCertificateInner<P>) -> Result<Self, Self::Error> {
         value.subject.validate()?;
