+++ src/constraints.rs
@@ -8,6 +8,7 @@ use x509_cert::name::Name;
 use crate::certs::capabilities::{Capabilities, KeyUsage};
 use crate::certs::idcsr::IdCsr;
 use crate::certs::SessionId;
+use crate::errors::base::ConstraintError;
 use crate::signature::Signature;
 use crate::Constrained;
 
@@ -23,7 +24,7 @@ impl Constrained for Name {
     ///     - uniqueIdentifier is the [SessionId] of the actor.
     /// - MAY have "organizational unit" attributes
     /// - MAY have other attributes, which might be ignored by other home servers and other clients.
-    fn validate(&self) -> Result<(), crate::ConstraintError> {
+    fn validate(&self) -> Result<(), ConstraintError> {
         // this code sucks. i couldn't think of a way to make it better though. sorry!
         let mut num_cn: u8 = 0;
         let mut num_dc: u8 = 0;
@@ -40,7 +41,7 @@ impl Constrained for Name {
                         if let Ok(value) = item.value.decode_as::<String>() {
                             SessionId::new_validated(value)?;
                         } else {
-                            return Err(crate::ConstraintError::Malformed(Some(
+                            return Err(ConstraintError::Malformed(Some(
                                 "Tried to decode SessionID as String and failed".to_string(),
                             )));
                         }
@@ -48,7 +49,7 @@ impl Constrained for Name {
                     "2.5.4.3" => {
                         num_cn += 1;
                         if num_cn > 1 {
-                            return Err(crate::ConstraintError::OutOfBounds {
+                            return Err(ConstraintError::OutOfBounds {
                                 lower: 1,
                                 upper: 1,
                                 actual: num_cn.to_string(),
@@ -62,7 +63,7 @@ impl Constrained for Name {
             }
         }
         if num_dc == 0 {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 1,
                 upper: u8::MAX as i32,
                 actual: "0".to_string(),
@@ -70,7 +71,7 @@ impl Constrained for Name {
             });
         }
         if num_uid > 1 {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 0,
                 upper: 1,
                 actual: num_uid.to_string(),
@@ -78,7 +79,7 @@ impl Constrained for Name {
             });
         }
         if num_unique_identifier > 1 {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 0,
                 upper: 1,
                 actual: num_unique_identifier.to_string(),
@@ -86,7 +87,7 @@ impl Constrained for Name {
             });
         }
         if num_unique_identifier > 0 && num_uid == 0 {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 1,
                 upper: 1,
                 actual: num_uid.to_string(),
@@ -97,7 +98,7 @@ impl Constrained for Name {
             });
         }
         if num_uid > 0 && num_unique_identifier == 0 {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 1,
                 upper: 1,
                 actual: num_unique_identifier.to_string(),
@@ -112,9 +113,9 @@ impl Constrained for Name {
 
 impl Constrained for SessionId {
     /// [SessionId] must be longer than 0 and not longer than 32 characters to be deemed valid.
-    fn validate(&self) -> Result<(), crate::ConstraintError> {
+    fn validate(&self) -> Result<(), ConstraintError> {
         if self.len() > Length::new(32) || self.len() == Length::ZERO {
-            return Err(crate::ConstraintError::OutOfBounds {
+            return Err(ConstraintError::OutOfBounds {
                 lower: 1,
                 upper: 32,
                 actual: self.len().to_string(),
@@ -126,7 +127,7 @@ impl Constrained for SessionId {
 }
 
 impl Constrained for Capabilities {
-    fn validate(&self) -> Result<(), crate::ConstraintError> {
+    fn validate(&self) -> Result<(), ConstraintError> {
         let is_ca = self.basic_constraints.ca;
 
         // Define the flags to check
@@ -162,14 +163,14 @@ impl Constrained for Capabilities {
         // Non-CAs must be able to sign their messages. Whether with or without non-repudiation
         // does not matter.
         if !is_ca && !can_sign && !can_commit_content {
-            return Err(crate::ConstraintError::Malformed(Some(
+            return Err(ConstraintError::Malformed(Some(
                 "Actors require signing capabilities, none found".to_string(),
             )));
         }
 
         // Certificates cannot be both non-repudiating and repudiating
         if can_sign && can_commit_content {
-            return Err(crate::ConstraintError::Malformed(Some(
+            return Err(ConstraintError::Malformed(Some(
                 "Cannot have both signing and non-repudiation signing capabilities".to_string(),
             )));
         }
@@ -178,12 +179,12 @@ impl Constrained for Capabilities {
         // true. Also, non-CAs are not allowed to have the KeyCertSign flag set to true.
         if is_ca || key_cert_sign {
             if !is_ca {
-                return Err(crate::ConstraintError::Malformed(Some(
+                return Err(ConstraintError::Malformed(Some(
                     "If KeyCertSign capability is wanted, CA flag must be true".to_string(),
                 )));
             }
             if !key_cert_sign {
-                return Err(crate::ConstraintError::Malformed(Some(
+                return Err(ConstraintError::Malformed(Some(
                     "CA must have KeyCertSign capability".to_string(),
                 )));
             }
@@ -193,7 +194,7 @@ impl Constrained for Capabilities {
         // See: <https://cryptography.io/en/latest/x509/reference/#cryptography.x509.KeyUsage.encipher_only>
         // See: <https://cryptography.io/en/latest/x509/reference/#cryptography.x509.KeyUsage.decipher_only>
         if (has_only_encipher || has_only_decipher) && !has_key_agreement {
-            Err(crate::ConstraintError::Malformed(Some(
+            Err(ConstraintError::Malformed(Some(
                 "KeyAgreement capability needs to be true to use OnlyEncipher or OnlyDecipher"
                     .to_string(),
             )))
@@ -204,7 +205,7 @@ impl Constrained for Capabilities {
 }
 
 impl<S: Signature> Constrained for IdCsr<S> {
-    fn validate(&self) -> Result<(), crate::ConstraintError> {
+    fn validate(&self) -> Result<(), ConstraintError> {
         self.inner_csr.capabilities.validate()?;
         self.inner_csr.subject.validate()?;
         Ok(())
